refactor(disposable): rename disposeable.js to disposable.js
The file is spelled incorrectly.
